Move getDeploymentByIdOrHost to /util/deploy
